Moving MRDWrapper to storage package
