+++ internal/storage/multi_range_downloader_wrapper.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package gcsx
+package storage
 
 import (
 	"bytes"
